Move the rule out to rules package
