+++ jetbrains-core/tst/software/aws/toolkits/jetbrains/utils/NotificationUtilsTest.kt
@@ -3,14 +3,11 @@
 
 package software.aws.toolkits.jetbrains.utils
 
-import com.intellij.notification.Notification
-import com.intellij.notification.Notifications
 import com.intellij.testFramework.ProjectRule
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.ExternalResource
-import java.util.concurrent.CopyOnWriteArrayList
+import software.aws.toolkits.jetbrains.utils.rules.NotificationListenerRule
 
 class NotificationUtilsTest {
 
@@ -31,19 +28,5 @@ class NotificationUtilsTest {
                 .startsWith("java.lang.NullPointerException")
                 .contains("NotificationUtilsTest.kt")
         }
-    }
-}
-
-class NotificationListenerRule(private val projectRule: ProjectRule) : ExternalResource() {
-    val notifications = CopyOnWriteArrayList<Notification>()
-
-    override fun before() {
-        with(projectRule.project.messageBus.connect()) {
-            setDefaultHandler { _, params ->
-                notifications.add(params[0] as Notification)
-            }
-            subscribe(Notifications.TOPIC)
-        }
-        notifications.clear()
     }
 }
